requirements.txt
torch
torchvision
torchaudio
transformers
datasets
peft
trl
accelerate
bitsandbytes
mlflow
mistralai
gspread
oauth2client
edsnlp
dspy
negspacy
matplotlib
scikit-learn
python-dotenv
presidio-analyzer 
presidio-anonymizer